third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/KeyframeEffect/composite-expected.txt

This is a testharness.js-based test.
FAIL Default value assert_equals: The default value should be replace expected (string) "replace" but got (undefined) undefined
PASS Change composite value
FAIL Unspecified keyframe composite value when setting effect composite anim.effect.getKeyframes is not a function
FAIL Specified keyframe composite value when setting effect composite anim.effect.getKeyframes is not a function
Harness: the test ran to completion.
